+++ AwsEncryptionSDK/runtimes/rust/examples/keyring/aws_kms_discovery_keyring_example.rs
