+++ AwsEncryptionSDK/runtimes/rust/examples/keyring/aws_kms_hierarchical/mod.rs
@@ -1,3 +1,3 @@
-pub mod aws_kms_hierarchical_keyring;
+pub mod aws_kms_hierarchical_keyring_example;
 pub mod create_branch_key_id;
 pub mod example_branch_key_id_supplier;
