+++ AwsEncryptionSDK/runtimes/rust/examples/keyring/aws_kms_keyring_example.rs
