+++ AwsEncryptionSDK/runtimes/rust/examples/keyring/aws_kms_mrk_discovery_keyring_example.rs
