+++ AwsEncryptionSDK/runtimes/rust/examples/keyring/aws_kms_rsa_keyring_example.rs
