+++ AwsEncryptionSDK/runtimes/rust/examples/keyring/mod.rs
@@ -1,13 +1,13 @@
-pub mod aws_kms_discovery_keyring;
-pub mod aws_kms_discovery_multi_keyring;
-pub mod aws_kms_keyring;
-pub mod aws_kms_mrk_discovery_keyring;
-pub mod aws_kms_mrk_discovery_multi_keyring;
-pub mod aws_kms_mrk_keyring;
-pub mod aws_kms_mrk_multi_keyring;
-pub mod aws_kms_multi_keyring;
-pub mod aws_kms_rsa_keyring;
-pub mod multi_keyring;
-pub mod raw_aes_keyring;
-pub mod raw_rsa_keyring;
+pub mod aws_kms_discovery_keyring_example;
+pub mod aws_kms_discovery_multi_keyring_example;
+pub mod aws_kms_keyring_example;
+pub mod aws_kms_mrk_discovery_keyring_example;
+pub mod aws_kms_mrk_discovery_multi_keyring_example;
+pub mod aws_kms_mrk_keyring_example;
+pub mod aws_kms_mrk_multi_keyring_example;
+pub mod aws_kms_multi_keyring_example;
+pub mod aws_kms_rsa_keyring_example;
+pub mod multi_keyring_example;
+pub mod raw_aes_keyring_example;
+pub mod raw_rsa_keyring_example;
 pub mod aws_kms_hierarchical;
