+++ AwsEncryptionSDK/runtimes/rust/examples/keyring/raw_aes_keyring_example.rs
